requirements.txt
asgiref==3.2.7
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==3.0.5
django-extensions==3.0.3
django-storages==1.9.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pycrypto==2.6.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.1.0